Added image upload and invoice generation feature

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const fileUtil = require('../util/file');
 const {validationResult} = require('express-validator/check');
 
 exports.getAddProduct = (req, res, next) => {
@@ -12,18 +13,30 @@ exports.getAddProduct = (req, res, next) => {
 };
 
 exports.postAddProduct = (req, res, next) => {
-    const {title, imageUrl, price, description} = req.body;
+    const {title, price, description} = req.body;
+    const image = req.file;
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Edit Product',
+            path: '/admin/edit-product',
+            editing: false,
+            product: {title, price, description},
+            hasError: true,
+            errorMessage: 'Image is not valid!',
+        });
+    }
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         return res.status(422).render('admin/edit-product', {
             pageTitle: 'Edit Product',
             path: '/admin/edit-product',
             editing: false,
-            product: {title, imageUrl, price, description},
+            product: {title, price, description},
             hasError: true,
             errorMessage: errors.array()[0].msg,
         });
     }
+    const imageUrl = image.path.replace("\\", "/");
     const product = new Product({title, price, description, imageUrl, userId: req.user});
     product.save()
         .then(_ => {
@@ -33,7 +46,7 @@ exports.postAddProduct = (req, res, next) => {
         .catch(err => {
             const error = new Error('Product creation failed!');
             error.httpStatusCode = 500;
-            next(error);
+            next(err);
         });
 };
 
@@ -60,11 +73,11 @@ exports.getEditProduct = (req, res, next) => {
 };
 
 exports.postEditProduct = (req, res, next) => {
-    const {productId, title, price, imageUrl, description} = req.body;
+    const {productId, title, price, description} = req.body;
     const updatedTitle = title;
     const updatedPrice = price;
     const updatedDescription = description;
-    const updatedImageUrl = imageUrl;
+    const image = req.file;
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         return res.status(422).render('admin/edit-product', {
@@ -73,7 +86,6 @@ exports.postEditProduct = (req, res, next) => {
             editing: false,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: productId,
@@ -87,7 +99,10 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDescription;
-            product.imageUrl = updatedImageUrl;
+            if (image) {
+                fileUtil.deleteFile(product.imageUrl);
+                product.imageUrl = image.path.replace("\\", "/");
+            }
             return product.save();
         })
         .then(_ => {
@@ -103,7 +118,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const {productId} = req.body;
-    Product.findByIdAndRemove(productId)
+    Product.findById(productId)
+        .then(product => {
+            if (!product) {
+                next(new Error('Product not found!'));
+            }
+            fileUtil.deleteFile(product.imageUrl);
+            return Product.findByIdAndRemove(productId);
+        })
         .then(_ => {
             console.log(`Product deleted!`);
             res.redirect('/admin/products');
